web/django_day7/django_day7_assignment/django_ws_7_3/requirements.txt
asgiref==3.7.2
Django==4.2.5
django-appconf==1.0.5
django-imagekit==4.1.0
pilkit==2.0
Pillow==9.5.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
